+++ ts/identity/key-derivation.ts
@@ -21,6 +21,7 @@ export function deriveMasterKeyPairFromSeedPhrase(seedPhrase: string) {
  * @returns {HDNode} - the derived child  generic signing key.
  *
 */
+
 export function deriveGenericSigningKeyPair(masterKeyPair: any) {
   return masterKeyPair.derivePath('m/73\'/0\'/0\'')
 }
@@ -33,6 +34,7 @@ export function deriveGenericSigningKeyPair(masterKeyPair: any) {
  * @returns {HDNode} - the derived Ethereum keypair.
  *
 */
+
 export function deriveEthereumKeyPair(masterKeyPair: any) {
   return masterKeyPair.derivePath('m/44\'/60\'/0\'/0/0')
 }
